monad-discovery: future proof name record structure
this change replaces single socket addr with the following fields:
- ip
- ports
- capabilities

ports is a collection, limited to 8 items, with port for tcp, udp.

capabilities is meant to be used for rolling upgrades. so for example
with direct socket, we will be able to use plain protocol selectively
and not necessarily at the same time when socket is enabled.

the other use case that i foresee in the future is the support of
authenticated protocol for udp connection, that would also benefit
from negotiated capability